+++ fiware-region-sanity-tests/tests/regions/test_waterford.py
@@ -24,8 +24,9 @@
 __author__ = 'jfernandez'
 
 
-from tests import fiware_region_with_networks_with_storage_tests
+from tests import fiware_region_with_networks_tests, fiware_region_object_storage_tests
 
 
-class TestSuite(fiware_region_with_networks_with_storage_tests.FiwareRegionWithNetworkWithStorageTest):
+class TestSuite(fiware_region_with_networks_tests.FiwareRegionWithNetworkTest,
+                fiware_region_object_storage_tests.FiwareRegionsObjectStorageTests):
     region_name = "Waterford"
